+++ plugins/plugin-chart-composed/test/index.test.ts
@@ -26,7 +26,7 @@ import ComposedChartPlugin from '../src';
  * treated correctly (e.g. formData from plugin controls
  * properly transform the data and/or any resulting props).
  */
-describe('@superset-ui/plugin-chart-composed', () => {
+describe('plugin-chart-composed', () => {
   it('exists', () => {
     expect(ComposedChartPlugin).toBeDefined();
   });