+++ plugins/plugin-chart-pivot-table/src/components/Layout.ts
@@ -1,7 +1,22 @@
-import { styled } from '@superset-ui/core';
-// @ts-ignore
-// eslint-disable-next-line import/no-unresolved
-import { supersetTheme } from '@superset-ui/style';
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import { styled, supersetTheme } from '@superset-ui/core';
 
 export const Grid = styled.div<{
   bordered?: boolean;
@@ -18,8 +33,7 @@ export const Grid = styled.div<{
   ${({ gridColumn }) => gridColumn && `grid-column: ${gridColumn};`}
   ${({ gridAutoFlow }) => gridAutoFlow && `grid-auto-flow: ${gridAutoFlow};`}
   ${({ gridAutoRows }) => gridAutoRows && `grid-auto-rows: ${gridAutoRows};`}
-  ${({ gridTemplateColumns }) =>
-    gridTemplateColumns && `grid-template-columns: ${gridTemplateColumns};`}
+  ${({ gridTemplateColumns }) => gridTemplateColumns && `grid-template-columns: ${gridTemplateColumns};`}
   ${({ gridTemplateRows }) => gridTemplateRows && `grid-template-rows: ${gridTemplateRows};`}
 `;
 
@@ -41,8 +55,9 @@ export const GridItem = styled(FillItem)<{
   gridRow?: string;
 }>`
   overflow: hidden;
-  ${({ bgLevel }) =>
-    bgLevel && `background-color: ${supersetTheme.colors.grayscale[`light${bgLevel}`]};`}
+  ${({ bgLevel, theme }) =>
+    bgLevel &&
+    `background-color: ${theme.colors.grayscale[`light${bgLevel}` as keyof typeof supersetTheme.colors.grayscale]};`}
   ${({ header }) => header && 'font-weight: bolder;'}
   ${({ bordered, hidden }) => bordered && !hidden && 'border: 1px solid black;'}
   ${({ gridColumn }) => gridColumn && `grid-column: ${gridColumn};`}