+++ plugins/plugin-chart-pivot-table/src/components/RowsHeader.tsx
@@ -1,6 +1,23 @@
-// @ts-ignore
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 import React, { FC } from 'react';
-import { styled, t } from '@superset-ui/core/lib';
+import { styled, t } from '@superset-ui/core';
 import { Grid, GridItem } from './Layout';
 import { ROW_HEIGHT, Unit } from '../plugin/utils';
 import HeadersOfHeader from './HeaderOfHeader';
@@ -27,13 +44,12 @@ const RowsHeader: FC<RowsHeaderProps<string, string>> = ({
   rowsFillData,
   showTotal,
 }) => {
-  const resultRowsData = rowsFillData;
   return (
     <div>
       <Grid
         withoutOverflow
         gridTemplateColumns={`repeat(${rows.length || 1}, max-content)`}
-        gridTemplateRows={`max-content max-content ${resultRowsData
+        gridTemplateRows={`max-content max-content ${rowsFillData
           .map(fillData => `${fillData ? ROW_HEIGHT : 0}`)
           .join(' ')}`}
         gridAutoFlow="column"